Move `fetch_commit_yaml_and_possibly_store` to `services.repository` (#604)
